Add some precondition checks
Simply make sure that the application has access to all the necessary
environment variables as well as the connections to AppSignal and
Graylog are up and running.

While this has been done with a try/rescue for convenience, it would be
much better to not have to crash/catch. Revisit this flow at some point.